Refactor code to improve code structure
