+++ requirements.txt
@@ -1,0 +1,3 @@
+requests==2.31.0
+beautifulsoup4==4.12.2
+python-dotenv==1.0.0
