+++ services/grade_service.py
@@ -1,0 +1,41 @@
+class GradeService:
+    def __init__(self, oge_scraper):
+        self.oge_scraper = oge_scraper
+
+    def getGrades(self, semester):
+        gradesPage = self.selectGradesSemester(semester)
+
+        # soup = BeautifulSoup(gradesPage, 'html.parser')
+
+        # To do later...
+
+        grades = []
+
+        return grades
+
+    def selectGradesSemester(self, semester):
+        headers = {
+            "Faces-Request": "partial/ajax",
+            "X-Requested-With": "XMLHttpRequest",
+            "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8"
+        }
+
+        data = {
+            "javax.faces.partial.ajax": "true",
+            "javax.faces.source": "mainBilanForm:j_id_15",
+            "javax.faces.partial.execute": "mainBilanForm:j_id_15",
+            "javax.faces.partial.render": "mainBilanForm",
+            "mainBilanForm:j_id_15": "mainBilanForm:j_id_15",
+            "i": str(int(semester) - 1),
+            "mainBilanForm:j_id_15_menuid": str(int(semester) - 1),
+            "mainBilanForm_SUBMIT": "1",
+            "javax.faces.ViewState": "0"
+        }
+
+        # Send the request
+        response = self.session.post(self.grades_url, headers=headers, data=data)
+
+        # Process the response to extract the real content
+        content = response.text.split("![CDATA[")[1].split("]]")[0]
+
+        return content
