+++ utils/data_processing.py
@@ -1,6 +1,5 @@
 from bs4 import BeautifulSoup
-from model.Absence import Absence
-from model.AbsenceProcessor import AbsenceProcessor
+from processors.absence_processor import AbsenceProcessor
 
 def create_absences(absencesPage):
 	soup = BeautifulSoup(absencesPage, 'html.parser')
